+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Feat: add `IGetValue.GetValue`.
 - Feat: add `IArrayState`, make `WatchArray` implement `IArrayState`.
 - Feat: add `Timer.UpdateWithDelta`.
+- Feat: accept callback in constructor of `Timer/RepeatedTimer`.
 - Fix: `AdvancedEvent` will call actions and once actions by the order they are added.
   - Previously all once actions will be called after repeated actions.
 - Fix: `AdvancedEvent` allow adding listeners or remove listeners during the invocation.